Remove: 12 Rules for Life
I've listen 45 minutes and I don't like the style of "writing", the proof for the concepts he explain seems anecdotal, he makes too many references to the christian bible (as if it were a source of truth) and makes some disturbing unfundamented distinctions between man and women that are troubling. It seems that he likes to polarize and sees the options as binary. I don't usually like this kind of thinking. I have more interesting reads in my list.